fix: dashboard layout 생성 및 Sidebar, Header 레이아웃 조정
